first commit
                 _Assignment 3_

Question 1:
Create a new column in the provided dataset. Column name should be num_of_cast. The assignment was already explained in class. You have to append or add a column in the dataset which contains the total number of cast in that particular movie or Tv-Show.

Question 2:
In the provided dataset count how many number of tv-shows and movies of each genre are there using group by or whichever other method you see fit.